+++ flow/activities/flowable_core.go
@@ -26,6 +26,13 @@ import (
 	"github.com/PeerDB-io/peer-flow/shared"
 )
 
+type PeerType string
+
+const (
+	Source      PeerType = "source"
+	Destination PeerType = "destination"
+)
+
 func heartbeatRoutine(
 	ctx context.Context,
 	message func() string,
@@ -527,4 +534,20 @@ func replicateXminPartition[TRead any, TWrite any, TSync connectors.QRepSyncConn
 	}
 
 	return currentSnapshotXmin, nil
+}
+
+func (a *FlowableActivity) getPeerNameForMirror(ctx context.Context, flowName string, peerType PeerType) (string, error) {
+	peerClause := "source_peer"
+	if peerType == Destination {
+		peerClause = "destination_peer"
+	}
+	q := fmt.Sprintf("SELECT p.name FROM flows f JOIN peers p ON f.%s = p.id WHERE f.name = $1;", peerClause)
+	var peerName string
+	err := a.CatalogPool.QueryRow(ctx, q, flowName).Scan(&peerName)
+	if err != nil {
+		slog.Error("failed to get peer name for flow", slog.String("flow_name", flowName), slog.Any("error", err))
+		return "", fmt.Errorf("failed to get peer name for flow %s: %w", flowName, err)
+	}
+
+	return peerName, nil
 }
